requirements.txt
asgiref==3.8.1
babel==2.17.0
Django==5.2
django-db-views==0.1.11
django-money==3.5.4
pillow==11.2.1
psycopg2==2.9.10
py-moneyed==3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
setuptools==79.0.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
